Add comments, remove dead code, isolate LeaderTenure
