+++ src/testnet/mod.rs
@@ -3,12 +3,14 @@ pub mod mem_pool;
 pub mod keychain;
 pub mod burnchain;
 pub mod node;
+pub mod tenure;
 
 pub use self::run_loop::{RunLoop};
 pub use self::mem_pool::{MemPool, MemPoolFS};
 pub use self::keychain::{Keychain};
 pub use self::node::{Node, SortitionedBlock};
 pub use self::burnchain::{BurnchainSimulator};
+pub use self::tenure::{LeaderTenure};
 
 pub struct Config {
     pub testnet_name: String,
